backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
openai>=1.40.0
websockets>=12.0
python-docx>=1.1.0
pymupdf>=1.24.0
pydub>=0.25.1
numpy>=1.26.0
aiofiles>=24.1.0
